migrated UserCollectionTests to Groovy in order to assert xml/html request/response efficiently
