--- src/test/resources/dav/users/usersHead_response.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" ?><D:error xmlns:cosmo="http://osafoundation.org/cosmo/DAV" xmlns:D="DAV:"><cosmo:internal-server-error></cosmo:internal-server-error></D:error>
