--- src/test/resources/dav/users/usersPost_response.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" ?><D:error xmlns:cosmo="http://osafoundation.org/cosmo/DAV" xmlns:D="DAV:"><cosmo:method-not-allowed>POST not allowed for a collection</cosmo:method-not-allowed></D:error>
