--- src/test/resources/dav/users/usersPropFindAllprop_request.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<D:propfind xmlns:D="DAV:">
-    <D:allprop />
-</D:propfind>
