--- src/test/resources/dav/users/usersPropFindProp_request.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<D:propfind xmlns:D="DAV:">
-
-    <D:prop>
-        <D:displayname>
-        </D:displayname>
-    </D:prop>
-
-</D:propfind>
