--- src/test/resources/dav/users/usersPropFindPropname_request.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<D:propfind xmlns:D="DAV:">
-    <D:propname />
-</D:propfind>
